website/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
google-generativeai
